update network calls to allow assistant requests to use the OAuth token if login flow was used, update headers applied to outgoing requests
