+++ internal/pkg/assistants/assistant_file_upload.go
@@ -19,7 +19,7 @@ func UploadAssistantFile(name string, filePath string) (*AssistantFileModel, err
 	resp, err := network.PostMultipartFormDataAndDecode[AssistantFileModel](
 		assistantDataUrl,
 		fmt.Sprintf(URL_ASSISTANT_FILE_UPLOAD, name),
-		true,
+		false,
 		filePath,
 	)
 	if err != nil {
