+++ internal/pkg/assistants/create.go
@@ -26,7 +26,7 @@ func CreateAssistant(name string) (*AssistantModel, error) {
 	resp, err := network.PostAndDecode[CreateAssistantRequest, AssistantModel](
 		assistantControlUrl,
 		URL_CREATE_ASSISTANT,
-		true,
+		false,
 		body,
 	)
 	if err != nil {
