+++ internal/pkg/assistants/delete.go
@@ -26,7 +26,7 @@ func DeleteAssistant(name string) (*DeleteAssistantResponse, error) {
 		assistantControlUrl,
 		pcio.Sprintf(URL_DELETE_ASSISTANT, name),
 		http.MethodDelete,
-		true,
+		false,
 	)
 	if err != nil {
 		return nil, err
