+++ internal/pkg/assistants/list.go
@@ -51,7 +51,7 @@ func ListAssistants() (*ListAssistantsResponse, error) {
 	resp, err := network.GetAndDecode[ListAssistantsResponse](
 		assistantControlUrl,
 		URL_LIST_ASSISTANTS,
-		true,
+		false,
 	)
 	if err != nil {
 		return nil, err
